requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
gradio==5.14.0
python-dotenv==1.0.1
pymongo==4.11
boto3==1.36.11
langchain_aws==0.2.12
langchain_mongodb==0.4.0
opencv-python==4.11.0.86
pysrt==1.1.2
scikit-image==0.25.1